fix audit tests
