+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/model/StudentExamAttributes.java
@@ -1,5 +1,7 @@
 package org.opentestsystem.rdw.ingest.processor.model;
 
+import java.time.LocalDate;
+
 /**
  * A student attributes at the time of the {@link Exam}
  */
@@ -18,6 +20,8 @@ public class StudentExamAttributes {
     private String languageCode;
     private String primDisabilityType;
     private Long examineeId;
+    private Integer elasId;
+    private LocalDate elasStartDate;
 
     public Integer getStudentId() {
         return studentId;
@@ -78,6 +82,14 @@ public class StudentExamAttributes {
         return examineeId;
     }
 
+    public Integer getElasId() {
+        return elasId;
+    }
+
+    public LocalDate getElasStartDate() {
+        return elasStartDate;
+    }
+
     public static Builder builder() {
         return new Builder();
     }
@@ -99,6 +111,8 @@ public class StudentExamAttributes {
         private String languageCode;
         private String primDisabilityType;
         private Long examineeId;
+        private Integer elasId;
+        private LocalDate elasStartDate;
 
 
         public StudentExamAttributes build() {
@@ -116,6 +130,8 @@ public class StudentExamAttributes {
             studentExamAttributes.languageCode = languageCode;
             studentExamAttributes.primDisabilityType = primDisabilityType;
             studentExamAttributes.examineeId = examineeId;
+            studentExamAttributes.elasId = elasId;
+            studentExamAttributes.elasStartDate = elasStartDate;
             return studentExamAttributes;
         }
 
@@ -181,6 +197,16 @@ public class StudentExamAttributes {
 
         public Builder examineeId(final Long examineeId) {
             this.examineeId = examineeId;
+            return this;
+        }
+
+        public Builder elasId(final Integer elasId) {
+            this.elasId = elasId;
+            return this;
+        }
+
+        public Builder elasStartDate(final LocalDate elasStartDate) {
+            this.elasStartDate = elasStartDate;
             return this;
         }
     }
