+++ exam-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/repository/impl/JdbcExamRepository.java
@@ -23,8 +23,12 @@ import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.Date;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.sql.Timestamp;
 import java.time.Instant;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
@@ -189,7 +193,7 @@ class JdbcExamRepository implements ExamRepository {
                 .studentId(row.getInt("student_id"))
                 .responsibleSchoolId(row.getInt("school_id"))
                 .ideaIndicator(row.getBoolean("iep"))
-                .lep(row.getBoolean("lep"))
+                .lep(getNullable(row, row.getBoolean("lep")))
                 .section504(getNullable(row, row.getBoolean("section504")))
                 .economicDisadvantage(row.getBoolean("economic_disadvantage"))
                 .migrantStatus(getNullable(row, row.getBoolean("migrant_status")))
@@ -198,6 +202,8 @@ class JdbcExamRepository implements ExamRepository {
                 .languageCode(row.getString("language_code"))
                 .primDisabilityType(row.getString("prim_disability_type"))
                 .examineeId(getNullable(row, row.getLong("examinee_id")))
+                .elasId(getNullable(row, row.getInt("elas_id")))
+                .elasStartDate(getNullableLocalDate(row, row.getDate("elas_start_at")))
                 .build();
         return Exam.builder()
                 .studentExamAttributes(attributes)
@@ -221,7 +227,7 @@ class JdbcExamRepository implements ExamRepository {
                 .handScoreProject(getNullable(row, row.getLong("hand_score_project")))
                 .contract(row.getString("contract"))
                 .testReason(row.getString("test_reason"))
-                .assessmentAdminStartedAt(row.getDate("assessment_admin_started_at") == null ? null : row.getDate("assessment_admin_started_at").toLocalDate())
+                .assessmentAdminStartedAt(getNullableLocalDate(row, row.getDate("assessment_admin_started_at")))
                 .startedAt(toInstantOrNull(row.getTimestamp("started_at")))
                 .forceSubmittedAt(toInstantOrNull(row.getTimestamp("force_submitted_at")))
                 .status(row.getString("status"))
@@ -297,6 +303,8 @@ class JdbcExamRepository implements ExamRepository {
                 .addValue("language_code", studentExamAttributes.getLanguageCode())
                 .addValue("prim_disability_type", studentExamAttributes.getPrimDisabilityType())
                 .addValue("examinee_id", studentExamAttributes.getExamineeId())
+                .addValue("elas_id", studentExamAttributes.getElasId())
+                .addValue("elas_start_at", studentExamAttributes.getElasStartDate())
                 .getValues();
     }
 
@@ -352,7 +360,9 @@ class JdbcExamRepository implements ExamRepository {
                 && Objects.equals(e1.getStudentExamAttributes().getT3ProgramType(), e2.getStudentExamAttributes().getT3ProgramType())
                 && Objects.equals(e1.getStudentExamAttributes().getLanguageCode(), e2.getStudentExamAttributes().getLanguageCode())
                 && Objects.equals(e1.getStudentExamAttributes().getPrimDisabilityType(), e2.getStudentExamAttributes().getPrimDisabilityType())
-                && Objects.equals(e1.getStudentExamAttributes().getExamineeId(), e2.getStudentExamAttributes().getExamineeId());
+                && Objects.equals(e1.getStudentExamAttributes().getExamineeId(), e2.getStudentExamAttributes().getExamineeId())
+                && Objects.equals(e1.getStudentExamAttributes().getElasId(), e2.getStudentExamAttributes().getElasId())
+                && Objects.equals(e1.getStudentExamAttributes().getElasStartDate(), e2.getStudentExamAttributes().getElasStartDate());
     }
 
     private boolean upsertExamClaims(final List<ExamClaim> submittedClaims, final long examId) {
@@ -671,5 +681,9 @@ class JdbcExamRepository implements ExamRepository {
 
     private static Instant toInstantOrNull(final Timestamp timestamp) {
         return timestamp == null ? null : timestamp.toInstant();
+    }
+
+    private static LocalDate getNullableLocalDate(final ResultSet rs, final Date localDate) throws SQLException {
+        return rs.wasNull() ? null : localDate.toLocalDate();
     }
 }
